requirements.txt
flask
line-bot-sdk
wheel
grpcio
google-cloud-firestore==2.0.2
flask_cors
line-bot-sdk
google-cloud-storage
pandas
gunicorn
google-cloud-logging
xlrd
openpyxl
firebase_admin
